+++ CHANGELOG.md
@@ -1,7 +1,41 @@
-## 0.1.1 (2025-06-02)
+## 0.4.0 (2025-06-03)
 
 ### Fix
 
-- trial
+- y
+- e
+- g
 
-## 0.1.0 (2025-06-02)
+## 0.3.0 (2025-06-03)
+
+### Fix
+
+- q
+- f
+- final
+- x
+
+## 0.1.1 (2025-06-03)
+
+### Fix
+
+- a
+
+## v1.0.1 (2025-06-03)
+
+## v0.1.0 (2025-06-03)
+
+### Fix
+
+- fix: main.py and legalclassifier.py updated. 20 legal classes are converted to object from hash map. Made agent instruction and descriptions configurable
+
+## v1.0.0 (2025-06-03)
+
+### Feat
+
+- add main logic
+
+### Fix
+
+- main.py and legalclassifier.py updated. 20 legal classes are converted to object from hash map. Made agent instruction and description configurable.
+- trial